move other svgs into folder
