+++ www/components/static-exporting/svg/Result.js
